+++ src/rer/solrpush/upgrades.py
@@ -1,6 +1,12 @@
 # -*- coding: utf-8 -*-
 from plone import api
 from plone.app.upgrade.utils import installOrReinstallProduct
+from rer.solrpush.utils.solr_common import init_solr_push
+from rer.solrpush.utils.solr_indexer import push_to_solr
+
+import logging
+
+logger = logging.getLogger(__name__)
 
 default_profile = "profile-rer.solrpush:default"
 
@@ -42,3 +48,25 @@ def to_1400(context):
     )
     update_registry(context)
     update_controlpanel(context)
+
+
+def to_1600(context):
+    # add new criteria
+    update_registry(context)
+    # reload schema
+    init_solr_push()
+    brains = api.content.find(portal_type=["File", "Image"])
+    tot = len(brains)
+    logger.info("Reindexing {} contents for updated mime_types.".format(tot))
+    i = 0
+
+    for brain in brains:
+        i += 1
+        if i % 500 == 0:
+            logger.info("[PROGRESS] - {}/{}".format(i, tot))
+        item = brain.getObject()
+        try:
+            push_to_solr(item)
+        except Exception:
+            # solr can't index it, pass
+            continue