+++ src/rer/solrpush/vocabularies.py
@@ -10,7 +10,10 @@ class AvailableSitesVocabularyFactory(object):
     @property
     def terms(self):
         solr_results = search(
-            query={"*": "*"}, fl="UID", facets=True, facet_fields="site_name"
+            query={"*": "*", "b_size": 1},
+            fl="UID",
+            facets=True,
+            facet_fields="site_name",
         )
         if isinstance(solr_results, dict) and solr_results.get("error", False):
             return []
@@ -29,4 +32,32 @@ class AvailableSitesVocabularyFactory(object):
         return SimpleVocabulary(self.terms)
 
 
+@implementer(IVocabularyFactory)
+class AvailableSubjectsVocabularyFactory(object):
+    @property
+    def terms(self):
+        solr_results = search(
+            query={"*": "*", "b_size": 1},
+            fl="UID",
+            facets=True,
+            facet_fields="Subject",
+        )
+        if isinstance(solr_results, dict) and solr_results.get("error", False):
+            return []
+        facets = solr_results.facets["facet_fields"].get("Subject", [])
+        if not facets:
+            return []
+        terms = []
+        for facet in facets:
+            for key in facet.keys():
+                terms.append(
+                    SimpleTerm(value=key, token=key.encode("utf-8"), title=key)
+                )
+        return terms
+
+    def __call__(self, context):
+        return SimpleVocabulary(self.terms)
+
+
 AvailableSitesVocabulary = AvailableSitesVocabularyFactory()
+AvailableSubjectsVocabulary = AvailableSubjectsVocabularyFactory()